Add always_step option to Newton solver
From updated README.md:

If the optional parameter `always_step = false` (the default), no Newton iteration is performed
if the solution is already converged for the supplied initial `x` values,  and the supplied initial `x` is returned
in the `zero` field of `SolverResults`. If `always_step = true`, a Newton step is performed
before testing for convergence.